+++ Assets/Scripts/CombatManager/ICombatant.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 public interface ICombatant
 {
     void PerformAction(List<ICombatant> enemies, ICombatReader combat, IEventRecipient<ICombatUpdateEvent> evRecipient);
-    void BeAttacked(int attackStat);
+    void BeAttacked(double attackStat);
     bool IsDead();
     ICombatAttributes Attributes { get; }
     ITurnProgress TurnProgress { get; }
